pyproject.toml
[project]
name = "fastapi-app"
version = "0.1.0"
description = "FastAPI project template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "alembic>=1.16.1",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "redis>=6.1.0",
    "ruff>=0.11.11",
    "slowapi>=0.1.9",
    "sqlmodel>=0.0.24",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = 88
